% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_registry.R
\name{.get_criterion_registry}
\alias{.get_criterion_registry}
\title{Internal Criterion Registry}
\usage{
.get_criterion_registry()
}
\value{
A named list of criterion configurations.
}
\description{
Internal Criterion Registry
}
\keyword{internal}
