% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{.get_model_registry}
\alias{.get_model_registry}
\title{Internal Model Registry}
\usage{
.get_model_registry()
}
\value{
A named list of model configurations.
}
\description{
Internal Model Registry
}
\keyword{internal}
